Extract the upgrade modal content into a shared UpgradePanel
The new upgrade modal had a much nicer pitch than the generic
UpsellContent block we were rendering inline for community users in
tabs and on the connection form. This pulls the modal's content out
into a standalone <upgrade-panel> component so we can render it both
inside the modal and inline anywhere we need to gate a feature.

- UpgradePanel takes an optional feature name and a `standalone` flag.
  Standalone adds card chrome (border, radius, modal-matching surface
  color) and centers itself; bare it just renders the content for the
  modal to wrap.
- UpgradeRequiredModal is now a thin shell around UpgradePanel plus the
  portal/close button.
- Replaced UpsellContent in TabImportTable, ImportExportDatabase,
  TabDatabaseBackup, TabPluginBase, TabPluginShell (non-AI-shell
  fallback), and ConnectionInterface, each with a feature name that
  matches its context (e.g. "Multi-Table Export", "Database Backup",
  the friendly connection type, the plugin tab title).
- Right-clicking a table and choosing Import from File now pops the
  upgrade modal for community users instead of opening a tab with the
  upsell - the gate is in beginImport so every entry point benefits.
- On the connection screen the panel renders as a sibling below the
  connection card rather than nested inside it, so it doesn't read as
  a card-in-a-card.
- Tightened the flex chain so the modal actually scrolls when its
  content exceeds the viewport, and the tab wrapper scrolls when the
  panel is taller than the tab area.

UpsellContent.vue and its SCSS are removed - nothing references them
anymore.
