+++ apps/studio/src/assets/styles/app/modals.scss
@@ -17,19 +17,33 @@
     .v--modal {
       overflow: hidden;
       border-radius: 12px;
+      max-height: calc(100vh - 40px);
+    }
+    .v--modal-box {
+      display: flex;
+      flex-direction: column;
     }
     .dialog-content.upgrade-modal-content {
       padding: 0;
       position: relative;
       display: flex;
       flex-direction: column;
-      max-height: calc(100vh - 40px);
-    }
-    .upgrade-modal-scroll {
-      overflow-y: auto;
-      padding: 22px 24px 4px;
       flex: 1 1 auto;
       min-height: 0;
+    }
+    .upgrade-panel {
+      display: flex;
+      flex-direction: column;
+      flex: 1 1 auto;
+      min-height: 0;
+    }
+    .upgrade-panel-scroll {
+      flex: 1 1 auto;
+      min-height: 0;
+      overflow-y: auto;
+    }
+    .upgrade-panel-header {
+      padding-right: 32px; // room for close button
     }
     .close-btn {
       position: absolute;
@@ -37,237 +51,6 @@
       right: 14px;
       z-index: 2;
       color: $text-light;
-    }
-
-    // Header
-    .upgrade-modal-header {
-      display: flex;
-      align-items: center;
-      gap: 14px;
-      padding-right: 32px; // room for close button
-
-      .bk-badge {
-        width: 38px;
-        height: 38px;
-        display: block;
-        flex-shrink: 0;
-      }
-      .title-block {
-        flex: 1;
-        min-width: 0;
-        &.triggered {
-          align-self: flex-start;
-        }
-        .eyebrow {
-          display: inline-flex;
-          align-items: center;
-          gap: 5px;
-          font-size: 11.5px;
-          color: $theme-primary;
-          text-transform: uppercase;
-          letter-spacing: 0.08em;
-          font-weight: 600;
-          margin-bottom: 2px;
-          .material-icons {
-            font-size: 13px;
-          }
-        }
-        .title {
-          margin: 0;
-          font-size: 18px;
-          font-weight: 500;
-          color: $text-dark;
-          letter-spacing: -0.1px;
-          line-height: 1.25;
-          padding-bottom: 0;
-        }
-      }
-      .indie-pill {
-        display: inline-flex;
-        align-items: center;
-        gap: 6px;
-        padding: 3px 8px;
-        font-size: 11px;
-        color: $text;
-        background: rgba($theme-base, 0.04);
-        border: 1px solid $border-color;
-        border-radius: 999px;
-        white-space: nowrap;
-        flex-shrink: 0;
-        .dot {
-          width: 6px;
-          height: 6px;
-          border-radius: 50%;
-          background: $brand-success;
-        }
-      }
-    }
-    .subtitle {
-      margin: 6px 0 0;
-      padding-left: 52px;
-      font-size: 13px;
-      color: $text-light;
-      line-height: 1.5;
-    }
-
-    // What you unlock
-    .unlock-section {
-      padding-top: 18px;
-      .section-label {
-        font-size: 11px;
-        color: $text-lighter;
-        text-transform: uppercase;
-        letter-spacing: 0.08em;
-        margin-bottom: 10px;
-        font-weight: 600;
-      }
-      .unlock-list {
-        margin: 0;
-        padding: 0;
-        list-style: none;
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-      }
-      .unlock-item {
-        display: flex;
-        align-items: flex-start;
-        gap: 11px;
-        list-style: none;
-        padding: 0;
-        &:before { content: none; }
-        .unlock-icon {
-          font-size: 18px;
-          margin-top: 1px;
-          flex-shrink: 0;
-        }
-        .unlock-text {
-          min-width: 0;
-          font-size: 13px;
-          line-height: 1.45;
-        }
-        .unlock-title {
-          color: $text-dark;
-          font-weight: 600;
-          margin-right: 6px;
-        }
-        .unlock-blurb {
-          color: $text-light;
-        }
-      }
-      .unlock-more {
-        display: flex;
-        align-items: center;
-        gap: 11px;
-        padding-left: 29px;
-        list-style: none;
-        &:before { content: none; }
-        a {
-          font-size: 12.5px;
-          color: $text-light;
-          text-decoration: none;
-          display: inline-flex;
-          align-items: center;
-          gap: 4px;
-          .link-emphasis {
-            color: $brand-secondary;
-            text-decoration: underline;
-            text-decoration-color: rgba($brand-secondary, 0.4);
-            margin-left: 2px;
-          }
-          .material-icons {
-            font-size: 14px;
-            color: $brand-secondary;
-          }
-        }
-      }
-    }
-
-    // Testimonial
-    .testimonial {
-      margin: 18px 0 0;
-      padding: 14px 16px;
-      background: rgba($theme-base, 0.03);
-      border: 1px solid $border-color;
-      border-radius: 8px;
-      display: flex;
-      gap: 14px;
-      align-items: flex-start;
-
-      .avatar {
-        width: 36px;
-        height: 36px;
-        flex-shrink: 0;
-        border-radius: 50%;
-        background: linear-gradient(135deg, rgba($theme-primary, 0.22), rgba($brand-secondary, 0.18));
-        border: 1px solid $border-color;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        color: $text-dark;
-        font-size: 14px;
-        font-weight: 600;
-        letter-spacing: 0.4px;
-      }
-      .testimonial-body {
-        flex: 1;
-        min-width: 0;
-      }
-      blockquote {
-        margin: 0;
-        font-size: 13px;
-        line-height: 1.55;
-        color: $text-dark;
-        font-weight: 400;
-      }
-      figcaption {
-        margin-top: 6px;
-        font-size: 11.5px;
-        color: $text-light;
-        display: flex;
-        align-items: center;
-        gap: 6px;
-        .author {
-          color: $text;
-          font-weight: 600;
-        }
-        .sep {
-          color: $text-lighter;
-        }
-      }
-    }
-
-    // CTAs — wraps the shared <UpsellButtons /> component
-    .cta-row {
-      margin-top: 18px;
-      padding: 16px 0 4px;
-      border-top: 1px solid $border-color;
-    }
-
-    // Lifetime license footer band
-    .lifetime-footer {
-      padding: 11px 24px;
-      background: rgba($theme-primary, 0.07);
-      border-top: 1px solid rgba($theme-primary, 0.18);
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 10px;
-      font-size: 12.5px;
-      color: $text;
-      flex-shrink: 0;
-
-      .material-icons {
-        font-size: 18px;
-        color: $theme-primary;
-      }
-      strong {
-        color: $text-dark;
-        font-weight: 600;
-      }
-      .asterisk {
-        color: $text-light;
-      }
     }
   }
