+++ apps/studio/src/assets/styles/components/_all.scss
@@ -7,5 +7,5 @@
 @import "dropzone.scss";
 @import "text-editor.scss";
 @import "json-viewer.scss";
-@import './upsell/upsell-content.scss';
+@import './upsell/upgrade-panel.scss';
 @import './upsell/ai-shell-upsell.scss';
