+++ apps/studio/src/components/TabPluginShell.vue
@@ -1,11 +1,10 @@
 <template>
   <div
     v-if="isCommunity && tab.context.pluginId.startsWith('bks-')"
-    class="tab-upsell-wrapper"
-    :class="{ 'tab-upsell-wrapper--ai-shell': isAiShellPlugin }"
+    :class="isAiShellPlugin ? 'tab-upsell-wrapper tab-upsell-wrapper--ai-shell' : 'upgrade-panel-tab-wrapper'"
   >
     <ai-shell-upsell v-if="isAiShellPlugin" />
-    <upsell-content v-else />
+    <upgrade-panel v-else :feature-name="tab.title || 'Plugins'" standalone />
   </div>
   <div v-else class="plugin-shell" ref="container" v-hotkey="keymap">
     <div class="top-panel" ref="topPanel">
@@ -90,7 +89,7 @@ import { TransportPluginTab } from "@/common/transport/TransportOpenTab";
 import IsolatedPluginView from "@/components/plugins/IsolatedPluginView.vue";
 import Vue from "vue";
 import { mapGetters } from "vuex";
-import UpsellContent from "@/components/upsell/UpsellContent.vue";
+import UpgradePanel from "@/components/upsell/UpgradePanel.vue";
 import AiShellUpsell from "@/components/upsell/AiShellUpsell.vue";
 import type { OnViewRequestListenerParams } from "@/services/plugin/types";
 import { RunQueryResponse } from "@beekeeperstudio/plugin"
@@ -106,7 +105,7 @@ export default Vue.extend({
     QueryEditorStatusBar,
     ErrorAlert,
     IsolatedPluginView,
-    UpsellContent,
+    UpgradePanel,
     AiShellUpsell,
   },
   props: {
